requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
pydantic-settings  
python-jose[cryptography]     
passlib[bcrypt]    
pydantic[email]             
PyJWT
python-multipart
alembic
pyotp
sentry-sdk[fastapi]
httpx
celery[asyncio]
rapidfuzz
billiard
tenacity
nest_asyncio
jinja2
weasyprint 
qrcode[pil]
asyncpg
fastapi-limiter
deep-translator
googletrans==4.0.0-rc1